fix(admin): trap focus in the initial-password modal
showInitialPassword opened the shared #modal element by hand: no Escape
key, no Tab trap, no focus restoration on close. Keyboard-only admins
had no reliable way to dismiss the dialog that reveals a freshly
generated credential.

Promote withModal from deck-sync.js to ui/shell.js (it now accepts
cancelLabel: null to hide the cancel button for one-button dialogs)
and route both the user-creation and reset-password reveals through it.
Behaviour for the deck-sync dialog is unchanged.
